+++ packages/core/src/components/navigation/Navbar/Navbar.tsx
@@ -1,29 +1,29 @@
-import { Suspense, useRef, useState } from 'react'
+import { useRef, useState } from 'react'
 
-import {
-  Icon as UIIcon,
-  IconButton as UIIconButton,
-  Navbar as UINavbar,
-  NavbarButtons as UINavbarButtons,
-  NavbarHeader as UINavbarHeader,
-  NavbarRow as UINavbarRow,
-  useScrollDirection,
-  useUI,
-} from '@faststore/ui'
+import { useUI, useScrollDirection, Icon as UIIcon } from '@faststore/ui'
 
 import { mark } from 'src/sdk/tests/mark'
 
-import CartToggle from 'src/components/cart/CartToggle'
 import type { SearchInputRef } from 'src/components/search/SearchInput'
 import SearchInput from 'src/components/search/SearchInput'
-
-import Link from 'src/components/ui/Link'
+import NavbarLinks from 'src/components/navigation/NavbarLinks'
+import NavbarSlider from 'src/components/navigation/NavbarSlider'
+import CartToggle from 'src/components/cart/CartToggle'
 import Logo from 'src/components/ui/Logo'
+import Link from 'src/components/ui/Link'
 import { ButtonSignIn } from 'src/components/ui/Button'
 
 import type { NavbarProps as SectionNavbarProps } from '../../sections/Navbar'
-import NavbarLinks from '../NavbarLinks'
-import NavbarSlider from '../NavbarSlider'
+
+import { Components, Props } from 'src/components/sections/Navbar/Overrides'
+
+const {
+  Navbar: NavbarWrapper,
+  NavbarHeader,
+  NavbarRow,
+  NavbarButtons,
+  IconButton,
+} = Components
 
 export interface NavbarProps {
   /**
@@ -77,7 +77,10 @@ function Navbar({
   home: { label: homeLabel },
   signIn: { button: signInButton },
   menu: {
-    icon: { icon: menuIcon, alt: menuIconAlt },
+    icon: {
+      icon: menuIcon,
+      alt: menuIconAlt = Props['IconButton']['aria-label'],
+    },
   },
 }: NavbarProps) {
   const scrollDirection = useScrollDirection()
@@ -91,16 +94,17 @@ function Navbar({
   }
 
   return (
-    <UINavbar scrollDirection={scrollDirection}>
-      <UINavbarHeader>
-        <UINavbarRow className="layout__content">
+    <NavbarWrapper scrollDirection={scrollDirection} {...Props['Navbar']}>
+      <NavbarHeader {...Props['NavbarHeader']}>
+        <NavbarRow className="layout__content" {...Props['NavbarRow']}>
           {!searchExpanded && (
             <>
-              <UIIconButton
+              <IconButton
                 data-fs-navbar-button-menu
                 onClick={openNavbar}
-                aria-label={menuIconAlt}
                 icon={<UIIcon name={menuIcon} width={32} height={32} />}
+                {...Props['IconButton']}
+                aria-label={menuIconAlt}
               />
               <Link
                 href="/"
@@ -116,9 +120,12 @@ function Navbar({
 
           <SearchInput sort={searchInput?.sort} />
 
-          <UINavbarButtons searchExpanded={searchExpanded}>
+          <NavbarButtons
+            searchExpanded={searchExpanded}
+            {...Props['NavbarButtons']}
+          >
             {searchExpanded && (
-              <UIIconButton
+              <IconButton
                 data-fs-button-collapse
                 aria-label="Collapse search bar"
                 icon={<UIIcon name="CaretLeft" width={32} height={32} />}
@@ -126,6 +133,7 @@ function Navbar({
                   setSearchExpanded(false)
                   searchMobileRef.current?.resetSearchInput()
                 }}
+                {...Props['IconButton']}
               />
             )}
 
@@ -143,9 +151,9 @@ function Navbar({
             <ButtonSignIn {...signInButton} />
 
             <CartToggle {...cart} />
-          </UINavbarButtons>
-        </UINavbarRow>
-      </UINavbarHeader>
+          </NavbarButtons>
+        </NavbarRow>
+      </NavbarHeader>
 
       <NavbarLinks links={links} region={region} className="hidden-mobile" />
 
@@ -158,7 +166,7 @@ function Navbar({
           region={region}
         />
       )}
-    </UINavbar>
+    </NavbarWrapper>
   )
 }