+++ packages/core/src/components/navigation/NavbarLinks/NavbarLinks.tsx
@@ -1,10 +1,6 @@
 import type { AnchorHTMLAttributes } from 'react'
 
-import {
-  NavbarLinks as UINavbarLinks,
-  NavbarLinksList as UINavbarLinksList,
-  NavbarLinksListItem as UINavbarLinksListItem,
-} from '@faststore/ui'
+import { NavbarLinksListItem as UINavbarLinksListItem } from '@faststore/ui'
 import type { NavbarLinksProps as UINavbarLinksProps } from '@faststore/ui'
 
 import Link from 'src/components/ui/Link'
@@ -12,6 +8,10 @@ import RegionButton from 'src/components/region/RegionButton'
 import type { NavbarProps } from 'src/components/navigation/Navbar'
 
 import { mark } from 'src/sdk/tests/mark'
+
+import { Components, Props } from 'src/components/sections/Navbar/Overrides'
+
+const { NavbarLinks: NavbarLinksWrapper, NavbarLinksList } = Components
 
 interface NavbarLinksProps extends UINavbarLinksProps {
   links: NavbarProps['links']
@@ -29,12 +29,12 @@ function NavbarLinks({
   ...otherProps
 }: NavbarLinksProps) {
   return (
-    <UINavbarLinks {...otherProps}>
+    <NavbarLinksWrapper {...otherProps} {...Props['NavbarLinks']}>
       <div className="layout__content">
         {shouldDisplayRegion && (
           <RegionButton icon={regionIcon} label={regionLabel} />
         )}
-        <UINavbarLinksList>
+        <NavbarLinksList {...Props['NavbarLinksList']}>
           {links.map(({ url, text }) => (
             <UINavbarLinksListItem key={text}>
               <Link variant="display" href={url} onClick={onClickLink}>
@@ -42,9 +42,9 @@ function NavbarLinks({
               </Link>
             </UINavbarLinksListItem>
           ))}
-        </UINavbarLinksList>
+        </NavbarLinksList>
       </div>
-    </UINavbarLinks>
+    </NavbarLinksWrapper>
   )
 }