+++ packages/core/src/components/navigation/NavbarSlider/NavbarSlider.tsx
@@ -1,22 +1,24 @@
-import {
-  NavbarSlider as UINavbarSlider,
-  NavbarSliderContent as UINavbarSliderContent,
-  NavbarSliderFooter as UINavbarSliderFooter,
-  NavbarSliderHeader as UINavbarSliderHeader,
-  useFadeEffect,
-  useUI,
-} from '@faststore/ui'
+import { useFadeEffect, useUI } from '@faststore/ui'
 import { Suspense } from 'react'
 
 import { ButtonSignIn, ButtonSignInFallback } from 'src/components/ui/Button'
 import Link from 'src/components/ui/Link'
+import NavbarLinks from 'src/components/navigation/NavbarLinks'
 import Logo from 'src/components/ui/Logo'
 import { mark } from 'src/sdk/tests/mark'
 
-import NavbarLinks from '../NavbarLinks'
 import type { NavbarProps } from '../Navbar'
 
 import styles from '../../sections/Navbar/section.module.scss'
+
+import { Components, Props } from 'src/components/sections/Navbar/Overrides'
+
+const {
+  NavbarSlider: NavbarSliderWrapper,
+  NavbarSliderHeader,
+  NavbarSliderContent,
+  NavbarSliderFooter,
+} = Components
 
 interface NavbarSliderProps {
   logo: NavbarProps['logo']
@@ -37,15 +39,16 @@ function NavbarSlider({
   const { fade, fadeOut } = useFadeEffect()
 
   return (
-    <UINavbarSlider
+    <NavbarSliderWrapper
       fade={fade}
       onDismiss={fadeOut}
       overlayProps={{
         className: `section ${styles.section} section-navbar-slider`,
       }}
       onTransitionEnd={() => fade === 'out' && closeNavbar()}
+      {...Props['NavbarSlider']}
     >
-      <UINavbarSliderHeader onClose={fadeOut}>
+      <NavbarSliderHeader onClose={fadeOut} {...Props['NavbarSliderHeader']}>
         <Link
           href="/"
           onClick={fadeOut}
@@ -55,14 +58,16 @@ function NavbarSlider({
         >
           <Logo {...logo} />
         </Link>
-      </UINavbarSliderHeader>
-      <UINavbarSliderContent>
+      </NavbarSliderHeader>
+      <NavbarSliderContent {...Props['NavbarSliderContent']}>
         <NavbarLinks onClickLink={fadeOut} links={links} region={region} />
-      </UINavbarSliderContent>
-      <UINavbarSliderFooter>
-        <ButtonSignIn {...signInButton} />
-      </UINavbarSliderFooter>
-    </UINavbarSlider>
+      </NavbarSliderContent>
+      <NavbarSliderFooter {...Props['NavbarSliderFooter']}>
+        <Suspense fallback={<ButtonSignInFallback />}>
+          <ButtonSignIn {...signInButton} />
+        </Suspense>
+      </NavbarSliderFooter>
+    </NavbarSliderWrapper>
   )
 }