+++ packages/core/src/components/sections/BannerNewsletter/BannerNewsletter.tsx
@@ -1,4 +1,7 @@
-import BannerText from 'src/components/ui/BannerText'
+import {
+  Banner as UIBanner,
+  BannerContent as UIBannerContent,
+} from '@faststore/ui'
 import Newsletter from '../Newsletter'
 import Section from '../Section'
 import styles from './section.module.scss'
@@ -9,14 +12,15 @@ function BannerNewsletter() {
       className={`${styles.section} section-banner-newsletter layout__content`}
     >
       <div data-fs-banner-newsletter>
-        <BannerText
-          title="Get to Know Our Next Release"
-          caption="Lorem ipsum dolor sit amet, consectetur adipiscing elit. Etiam elit nisi, vehicula in turpis sit amet, posuere aliquam nisl. "
-          actionLabel="Shop Now"
-          actionPath="/"
-          variant="secondary"
-          colorVariant="light"
-        />
+        <UIBanner variant="secondary" colorVariant="light">
+          <UIBannerContent
+            title="Get to Know Our Next Release"
+            caption="Lorem ipsum dolor sit amet, consectetur adipiscing elit. Etiam elit nisi, vehicula in turpis sit amet, posuere aliquam nisl. "
+            link="/"
+            linkText="Shop Now"
+          />
+        </UIBanner>
+
         <Newsletter
           title="Get News and Special Offers!"
           description="Receive our news and promotions in advance. Enjoy and get 10% off your first purchase. For more information click here."