+++ packages/core/src/components/sections/Navbar/Navbar.tsx
@@ -1,8 +1,8 @@
-import UINavbar from '../../navigation/Navbar'
-
 import Section from '../Section'
 
 import styles from './section.module.scss'
+
+import Navbar from 'src/components/navigation/Navbar'
 
 type PageLinks = {
   url: string
@@ -51,7 +51,7 @@ export interface NavbarProps {
   }
 }
 
-function Navbar({
+function NavbarSection({
   logo,
   searchInput,
   cartIcon,
@@ -68,8 +68,8 @@ function Navbar({
   },
 }: NavbarProps) {
   return (
-    <Section className={`${styles.section} section-navbar`}>
-      <UINavbar
+    <Section as="header" className={`${styles.section} section-navbar`}>
+      <Navbar
         home={home}
         menu={menu}
         logo={logo}
@@ -87,4 +87,4 @@ function Navbar({
   )
 }
 
-export default Navbar
+export default NavbarSection