+++ packages/core/src/customizations/components/overrides/WebFonts.tsx
@@ -1,0 +1,9 @@
+import { default as CustomWebFonts } from 'src/customizations/fonts/WebFonts'
+
+const overrides = {
+  components: {
+    WebFonts: CustomWebFonts,
+  },
+}
+
+export default overrides