+++ packages/core/src/typings/overrides.d.ts
@@ -1,18 +1,55 @@
 export const SECTIONS = {
-    ProductDetails: { components: ["Price"] },
-    ProductShelf: { components: ["ProductCard"] }
-  } as const;
-  
-  // export type ComponentOrProps =
-  //   | { Component: React.ElementType }
-  //   | { props: Record<string, unknown> };
-  
-  export type SectionOverride = {
-    [K in keyof typeof SECTIONS]: {
-      name: K;
-      components: {
-        [ComponentKey in typeof SECTIONS[K]["components"][number]]?: React.ElementType;
-      };
-    };
-  };
-  
+  Hero: {
+    components: ['Hero', 'HeroImage', 'HeroHeader'],
+  },
+  BannerText: {
+    components: ['Banner', 'BannerContent'],
+  },
+  ProductDetails: {
+    components: [
+      'ProductTitle',
+      'DiscountBadge',
+      'BuyButton',
+      'SkuSelector',
+      'ShippingSimulation',
+    ],
+  },
+  ProductShelf: {
+    components: [
+      'ProductShelf',
+      '__experimentalCarousel',
+      '__experimentalProductCard',
+    ],
+  },
+  Navbar: {
+    components: [
+      'Navbar',
+      'NavbarLinks',
+      'NavbarLinksList',
+      'NavbarSlider',
+      'NavbarSliderHeader',
+      'NavbarSliderContent',
+      'NavbarSliderFooter',
+      'NavbarHeader',
+      'NavbarRow',
+      'NavbarButtons',
+      'IconButton',
+    ],
+  },
+  Breadcrumb: {
+    components: ['Breadcrumb', 'Icon'],
+  },
+} as const
+
+// export type ComponentOrProps =
+//   | { Component: React.ElementType }
+//   | { props: Record<string, unknown> };
+
+export type SectionOverride = {
+  [K in keyof typeof SECTIONS]: {
+    name: K
+    components: {
+      [ComponentKey in (typeof SECTIONS)[K]['components'][number]]?: any
+    }
+  }
+}