+++ packages/lighthouse/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 2.0.173-alpha.0 (2023-06-01)
+
+**Note:** Version bump only for package @faststore/lighthouse
+
 ## [2.0.167-alpha.0](https://github.com/vtex/faststore/compare/v2.0.166-alpha.0...v2.0.167-alpha.0) (2023-05-29)
 
 ### Chores